Implementada a dependencia da JImageHash, e incluido o controller!
