+++ src/main/java/com/exemplo/imagem/database/Controller/ImageController.java
@@ -1,0 +1,39 @@
+package com.exemplo.imagem.database.Controller;
+
+import java.io.IOException;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
+
+import com.exemplo.imagem.database.service.ImageService;
+
+@RestController
+@RequestMapping("/image")
+public class ImageController {
+
+    @Autowired
+	private ImageService service;
+
+	@PostMapping
+	public ResponseEntity<?> uploadImage(@RequestParam("image") MultipartFile file) throws IOException {
+		String uploadImage = service.uploadImage(file);
+		return ResponseEntity.status(HttpStatus.OK).body(uploadImage);
+	}
+
+	@GetMapping("/{fileName}")
+	public ResponseEntity<?> downloadImage(@PathVariable String fileName) {
+		byte[] imageData = service.downloadImages(fileName);
+		return ResponseEntity.status(HttpStatus.OK).contentType(MediaType.valueOf("image/jpeg")).body(imageData);
+
+	}   
+    
+}
